#37 - refactored gitstrap.go
