+++ go.mod
@@ -2,7 +2,6 @@ module github.com/g4s8/gitstrap
 
 require (
 	github.com/g4s8/gopwd v0.0.0-20191109142035-f7311044bc0a
-	github.com/g4s8/gitstrap v0.0.0-20191003172754-76ad177c10f3
 	github.com/google/go-github v17.0.0+incompatible
 	github.com/google/go-querystring v1.0.0 // indirect
 	github.com/stretchr/testify v1.4.0
